+++ plone/app/widgets/utils.py
@@ -4,6 +4,7 @@ from Acquisition import aq_base
 from Acquisition import aq_parent
 from datetime import datetime
 from OFS.interfaces import IFolder
+from OFS.interfaces import ISimpleItem
 from plone.app.layout.navigation.root import getNavigationRootObject
 from Products.CMFCore.interfaces import ISiteRoot
 from Products.CMFCore.utils import getToolByName
@@ -194,6 +195,12 @@ def get_tinymce_options(context, field, request):
     """
     options = {}
     try:
+
+        if IForm.providedBy(context):
+            context = context.context
+        elif not ISimpleItem.providedBy(context):
+            context = getSite()
+
         pattern_options = getMultiAdapter(
             (context, request, field),
             name="plone_settings").tinymce()['data-pat-tinymce']